requirements.txt
langgraph>=0.2.0
langgraph-sdk>=0.1.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=9.0.0
structlog>=24.4.0